src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "typer>=0.20.0",
]
